Extract HighLine::BuiltinStyles as a module
